[refactor] moved emailService to proper place
